Refatoracao de funcoes para serem organizadas em diretorios especificos
